+++ gantry/__main__.py
@@ -314,6 +314,7 @@ def follow(experiment: str):
 @click.option(
     "-b", "--budget", type=str, help="""The budget account to associate with the experiment."""
 )
+@click.option("--stop-preemptible", is_flag=True, help="""Stop all preemptible on the cluster.""")
 def run(
     arg: Tuple[str, ...],
     name: Optional[str] = None,
@@ -349,6 +350,7 @@ def run(
     host_networking: bool = False,
     mount: Optional[Tuple[str, ...]] = None,
     budget: Optional[str] = None,
+    stop_preemptible: bool = False,
 ):
     """
     Run an experiment on Beaker.
@@ -547,6 +549,21 @@ def run(
 
     print(f"Experiment submitted, see progress at {beaker.experiment.url(experiment)}")
 
+    if stop_preemptible:
+        if priority == Priority.preemptible:
+            print_stderr("[yellow]You cannot preempt other jobs when your job is preemptible.[/]")
+        elif not cluster:
+            print_stderr("[yellow]Preempt jobs requires specifying a cluster.[/]")
+        elif len(cluster) > 1:
+            print_stderr("[yellow]Preempt jobs requires specifying a single cluster.[/]")
+        elif not dry_run:
+            print(f"Preempting jobs on cluster {cluster[0]}...")
+            preempted = beaker.cluster.preempt_jobs(cluster[0])
+            if preempted:
+                print(f"Preempted {len(preempted)} jobs on cluster {cluster[0]}")
+            else:
+                print("No jobs to preempt")
+
     # Can return right away if timeout is 0.
     if timeout == 0:
         return
@@ -691,6 +708,28 @@ def cluster_util(cluster: str):
             f"    GPUs free: [{'green' if node.free.gpu_count else 'red'}]"
             f"{node.free.gpu_count or 0} / {node.limits.gpu_count}[/] {node.free.gpu_type or ''}\n"
         )
+
+
+@cluster.command(name="allow-preemptible", **_CLICK_COMMAND_DEFAULTS)  # type: ignore
+@click.argument("cluster", nargs=1, required=True, type=str)
+def cluster_allow_preemptible(cluster: str):
+    """
+    Allow preemptible jobs on the cluster.
+    """
+    beaker = Beaker.from_env(session=True)
+    beaker.cluster.update(cluster, allow_preemptible=True)
+    print("[green]\N{check mark} Preemptible jobs allowed[/]")
+
+
+@cluster.command(name="disallow-preemptible", **_CLICK_COMMAND_DEFAULTS)  # type: ignore
+@click.argument("cluster", nargs=1, required=True, type=str)
+def cluster_disallow_preemptible(cluster: str):
+    """
+    Disallow preemptible jobs on the cluster.
+    """
+    beaker = Beaker.from_env(session=True)
+    beaker.cluster.update(cluster, allow_preemptible=False)
+    print("[yellow]\N{ballot x} Preemptible jobs disallowed[/]")
 
 
 if __name__ == "__main__":